+++ PaxstoreOpenApi/Paxstore/OpenApi/Model/PushHistory/OptimizedParamPushHistoryPageResponse.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Paxstore.OpenApi.Base.Dto;
+
+namespace Paxstore.OpenApi.Model.PushHistory
+{
+    public class OptimizedParamPushHistoryPageResponse: PageResponse<OptimizedParamPushHistory>
+    {
+    }
+}