+++ PaxstoreOpenApi/Paxstore/OpenApi/Model/TerminalVariable/TerminalParameterVariable.cs
@@ -11,6 +11,7 @@ namespace Paxstore.OpenApi.Model
         public long ID { get; set; }
         public Nullable<long> AppId { get; set; }
         public string AppPackageName { get; set; }
+        public string Type { get; set; }
         public string AppName { get; set; }
         public string Key { get; set; }
         public string Value { get; set; }