+++ PaxstoreOpenApi/Properties/AssemblyInfo.cs
@@ -6,5 +6,5 @@ using System.Text;
 using System.Threading.Tasks;
 
 
-[assembly: AssemblyVersion("7.4.0")]
-[assembly: AssemblyFileVersion("7.4.0")]
+[assembly: AssemblyVersion("8.1.0")]
+[assembly: AssemblyFileVersion("8.1.0")]